Initial commit
Creates the initial "mostly working" port of flamegraph.pl from
brendangregg/FlameGraph project.

Expect in the future:

- A proper test suite
- Performance improvements
- A v1.0 release on rubygems

NOTE:  This is not and equivalent patch to master's "Initial commit".
This commit was rebased multiple times to allow the master branch to
move forward with non-performance changes, while allowing this branch to
be a chronological logging of the changes I did to improve the
performance of this library after initially porting it from perl.  There
might be changes inspired from this branch that already exist in `master`,
or will be ported as part of this change, but this branch will NEVER be
merged into master or is intended to be as-is.