[PERFORMANCE] Single Data#parse_line regexp
Changes the regexp logic to instead search from the end of the string
instead of the front.  This avoids the need for the lazy `\s+?`, which
is tough to work with variable matches preceding it, and allows for
capturing just the samples1 and samples2 values.

The `prematch` capture ($`) which is already being defined anyway, can
then hold the stack data after we capture the other matches.  The new
regexp will then capture samples1, and samples2 (if it exists).

The performance improvement is minor, but reduces the overall logic of
the method to make it a worth while improvement, while not really
increasing the complexity of the regexp too much.