Add rakelib/git.rake
Add in a couple of git check tasks in charge of confirming that their
isn't dirty changes in the project, checking the tag is proper for a
release, and performing a `git tag` for a release.

Mostly pulled from the `bunlder/gem_tasks`, but uses more "Rake-isms"
instead of pulling in `Bundler::UI` code to do it.