[Rakefile] Add global release tasks
These two tasks will be in charge of creating a release on all
configured platforms.  For now, that is:

- github.com
- rubygems.org
- rubygems.pkg.github.com

The `release:manual` is intended for when releasing through GitHub
actions is down, but otherwise, it is assumed that just a `rake release`
will be sufficient to trigger the automated release.

(The release on GH actions will most likely be triggered via a release
event on `github.com`.  AKA:  When a commit is tagged, which will
trigger the gem package builds)