[Rakefile][rakelib] Use $LOAD_PATH for requires
While `rakelib/` will load any `*.rake` files from that directory
relative to the `Rakefile`, it doesn't update the `$LOAD_PATH` in doing
so.

This would be fine if the tasks in `rakelib/` were mostly `FileTasks`
and such, but because of the gem release process that is being built
have some extra complexities that require the use of libraries like
`net/http` and `git` helpers, they are better suited to be wrapped in a
class.

The `$LOAD_PATH` updates allow the individual tasks in the `rakelib/` to
be loaded piecemeal (if necessary), but allow the `require` statements
to look a little more readable to the end user.